backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
python-dotenv
httpx
imageio-ffmpeg
google-genai
